added test before gpg encrypt

+++ src/controllers/instance/backup.php
@@ -137,13 +137,19 @@ function instance_backup(array $data): array {
     exec("rm -rf $tmp_backup_dir");
 
     if($data['encrypt']) {
-        // Encrypt backup
-        exec("gpg --trust-model always --output $backup_file.gpg --encrypt --recipient $gpg_name $backup_file");
+		
+		exec("gpg --list-keys " . escapeshellarg($gpg_name) . " 2>&1", $output, $return_var);
 
-        // Remove non-encrypted backup (keep only crypted one)
-        unlink($backup_file);
+		// gpgp key found
+		if($return_var === 0) {
+			// Encrypt backup
+			exec("gpg --trust-model always --output $backup_file.gpg --encrypt --recipient $gpg_name $backup_file");
 
-        $backup_file = $backup_file.'.gpg';
+			$backup_file = $backup_file.'.gpg';
+			
+			// Remove non-encrypted backup (keep only crypted one)
+			unlink($backup_file);
+		}
     }
 
     return [
